+++ package/cpp/rnwgpu/api/GPUTexture.h
@@ -1,10 +1,5 @@
 #pragma once
 
-#include <future>
-#include <memory>
-#include <string>
-#include <vector>
-
 #include "Unions.h"
 
 #include "RNFHybridObject.h"
@@ -33,14 +28,14 @@ public:
       std::optional<std::shared_ptr<GPUTextureViewDescriptor>> descriptor);
   void destroy();
 
-  double getWidth();
-  double getHeight();
-  double getDepthOrArrayLayers();
-  double getMipLevelCount();
-  double getSampleCount();
+  uint32_t getWidth();
+  uint32_t getHeight();
+  uint32_t getDepthOrArrayLayers();
+  uint32_t getMipLevelCount();
+  uint32_t getSampleCount();
   wgpu::TextureDimension getDimension();
   wgpu::TextureFormat getFormat();
-  double getUsage();
+  uint32_t getUsage();
 
   std::string getLabel() { return _label; }
